backend/fastapi/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
pandas
numpy
scikit-learn
joblib
psycopg2-binary
python-dotenv